Aplicando correçoes de campos

+++ HTMLCSSModulo3/django/projeto1/app1/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class User(models.Model):
     nome = models.CharField('nome', max_length=30)
-    telefone = models.IntegerField('Telefone')
+    telefone = models.BigIntegerField('Telefone')
     email = models.CharField('e-mail', max_length=30)
     
     def __str__(self):
